media: ov5640: Change downsizing classification
Rather than having a single enum that describes the methods the
ov5640 is using to downsize an image, it makes more sense to
consider each method individually because their use each has
different ramifications. The methods are binning and scaling.

Since the binning process is much closer to the sensor, it seems
more likely the exposure calculation is needed after changing that
method.

Signed-off-by: Sam Bobrowicz <sam@elite-embedded.com>