media: ov5640: Add support for MIPI mode
This commit fixes MIPI mode, which appears to have been broken
since this driver was published. Most of the fixes are related
to the internally generated clocks. Comments are included to
document these changes.

Also, the PCLK period needed to be set in order for the MIPI
timing parameters to be properly met. This sets the UI value as
described in the OV5640 datasheet.

Signed-off-by: Sam Bobrowicz <sam@elite-embedded.com>